+++ packages/docusaurus-plugin-openapi-docs/src/types.ts
@@ -6,7 +6,7 @@
  * ========================================================================== */
 
 import { SidebarItemDoc } from "@docusaurus/plugin-content-docs/src/sidebars/types";
-import type Request from "postman-collection";
+import Request from "postman-collection";
 
 import {
   InfoObject,
@@ -118,6 +118,7 @@ export interface ApiMetadataBase {
   frontMatter: Record<string, unknown>;
   method?: string;
   path?: string;
+  position?: number;
 }
 
 export interface ApiPageMetadata extends ApiMetadataBase {