+++ packages/docusaurus-theme-openapi-docs/src/theme/ApiExplorer/Request/index.tsx
@@ -26,7 +26,7 @@ import Server from "@theme/ApiExplorer/Server";
 import { useTypedDispatch, useTypedSelector } from "@theme/ApiItem/hooks";
 import { ParameterObject } from "docusaurus-plugin-openapi-docs/src/openapi/types";
 import { ApiItem } from "docusaurus-plugin-openapi-docs/src/types";
-import sdk from "postman-collection";
+import * as sdk from "postman-collection";
 import { FormProvider, useForm } from "react-hook-form";
 
 import makeRequest from "./makeRequest";
@@ -126,7 +126,7 @@ function Request({ item }: { item: ApiItem }) {
       } else {
         await handleResponse(res);
       }
-    } catch (e: any) {
+    } catch (e) {
       console.log(e);
       dispatch(setResponse("Connection failed"));
       dispatch(clearCode());